+++ include/ellis/stream/cpp_input_stream.hpp
@@ -27,9 +27,9 @@ class cpp_input_stream : public sync_input_stream {
 public:
   cpp_input_stream(std::istream &is);
 
-  bool next_input_buf(byte **buf, int *bytecount) override;
+  bool next_input_buf(byte **buf, size_t *bytecount) override;
 
-  void put_back(int bytecount) override;
+  void put_back(size_t bytecount) override;
 
   std::unique_ptr<err> get_input_error() override;
 };
