+++ include/ellis/stream/fd_input_stream.hpp
@@ -24,8 +24,8 @@ class fd_input_stream : public sync_input_stream {
   int m_valid = 0;
 public:
   fd_input_stream(int fd);
-  bool next_input_buf(byte **buf, int *bytecount) override;
-  void put_back(int bytecount) override;
+  bool next_input_buf(byte **buf, size_t *bytecount) override;
+  void put_back(size_t bytecount) override;
   std::unique_ptr<err> get_input_error() override;
 };
